Initialize elements of d_pageidx to 0 if maxpage=1.
If there is only one page in the neighbor list (maxpage=1) then we need to initialise the elements of d_pageidx to 0 (since we don't bother running the <decode_neighlist_p1> kernel).
If we don't do this then we'll get garbage values and undefined (probably segfault) during the run stage.