TpA implementations (CL and CUDA) use shared memory to preload particle-i data.
Each block is given a set (one for each RO per-particle dataset) of fixed-size arrays in local memory.
These are preloaded from global memory, taking care to do so in a coalesced manner.
If there are redundant threads (in the last block) then we consider two approaches:
1) An outer rangecheck:
if (idx < N) preload()
2) Preload addresses modulo N
We need to benchmark to work out which is better.