Preloading for CL and CUDA BPA decompositions.
But not yet using data.
This is slightly dangerous because we assume the neighbor list has enough padding to allow each block to load BLOCK_SIZE*d contiguous elements (in multiple passes as jjbase increments). For the last block this could cause a segfault but we know that the neighbor list adds enough padding for this to be safe enough. Still... should come back to this.