Allow CUDA BPA to work with 2D grids.
Necessary for large problem sizes.
The BPA decomposition assigns one block per particle.
Assuming a 1D grid limits us to working with particle problem sizes =< 65535 (on a tesla).
We now switch to using a 2D grid if necessary.
This allows us to work up to problem sizes of =< 65535**2 (2**32) particles.

If necessary, we can move into 3D grids.
For now, we assert that the problem size is within bounds (and abort otherwise).

The macro MAX_GRID_DIM might be hardware specific...